==> R/1.preprocess-eda.Rmd <==
---
title: "K Means Method"
author: "Roghan"
date: "20 March 2023"
output: word_document
---

```{r}
library(tidyverse)
library(dplyr)
library(GGally)
library(ggplot2)
```

***Read Data

```{r}
df <- read.csv("./data-1.csv")
df %>% head
```

***Drop Unnecessary Columns

```{r}
df <- select(df, -c("Start.time","Completion.time","Email", "Name"))
df
```

***Rename Columns

```{r}
colnames(df) <- c('id', 'sex', 'yr', 'age', 'cgpa', 'bdevice', 'adevice', 'btime', 'atime', 'bdevstudy',      'adevstudy', 'bconc', 'aconc', 'bbed', 'abed', 'baffbed', 'aaffbed',
    'btire', 'atire', 'socdist', 'isolate', 'personality', 'lazy', 'boredom', 'psycho',                       'afford','unhealthy', 'covidstress', 'devicegpa', 'assgn', 'f2f', 'quiz')
```
```{r}
df
```

***Drop Null Rows

```{r}
df[complete.cases(df), ]
```

Therefore, No null values to be dropped

***Normalizing CGPA and Answer values

```{r}
for (row in 1:nrow(df)) {
    cgpa <- df[row, "cgpa"]
    df[row, "cgpa"] = trimws(unlist(strsplit(cgpa, "/"))[1])
}

```

```{r}
cols <- colnames(df)[10:ncol(df)]
for(i in cols){
  
  for (row in 1:nrow(df)) {
    v <- trimws(df[row, i])
    
    if(v=="Strongly Disagree"){
      curr= 0
    }
    else if (v=="Disagree"){
      curr= 1
    }
    else if (v=="Uncertain"){
      curr= 2
    }
    else if (v=="Agree"){
      curr= 3
    }
    else if (v=="Strongly Agree"){
      curr= 4
    }
    df[row, i] = curr
  }
}
```

```{r}
df$btime[df$btime=="+12"]<-"12+"
df$atime[df$atime=="+12"]<-"12+"

```

```{r}
df
```


***Analyzing Device used

```{r}
plot= data.frame(device= df$bdevice, sex= df$sex, time= c("Before COVID-19"))
plot= rbind(plot, data.frame(device= df$adevice, sex= df$sex, time= c("After COVID-19")))

```

```{r}
plot %>%
  mutate(time = factor(time, levels = c("Before COVID-19","After COVID-19"))) %>%
  ggplot(aes(x=device, fill= time)) +
  geom_bar(position="dodge") +
  labs(title="Change in the use of device before and after COVID-19", x= "Device Used") + 
  theme_bw()
```
We can infer from the above barplot that the use of Laptops as well as Personal computers has increased. We'll be assuming that this increase is due to the academic responsibilities of the students. 

The change in the use of Ipad has remained more or less the same.  

There is a drastic decrease in the use of Mobile phones and Other devices, since they had to switch over to Laptops/Computers.


***Analyzing Time Spent On device for Studies


```{r}
plot2= data.frame(devtime= df$btime, time= c("Before COVID-19"))
plot2= rbind(plot2, data.frame(devtime= df$atime, time= c("After COVID-19")))
```
```{r}
plot2 %>%
  mutate(time = factor(time, levels = c("Before COVID-19","After COVID-19"))) %>%
  mutate(devtime = factor(devtime, levels = c("1-3 ","3-6","6-9","9-12","12+"))) %>%
  ggplot(aes(x=devtime, fill= time)) +
  geom_bar(position="dodge") +
  labs(title="Change in the usage time of the device for studies before and after COVID-19", x= "Amount of time spent") + 
  theme_bw()
```
We can infer from the above barplot that the usage time of the devices for studies has drastically increased from the 1-3 hr Range to 3-6/6-9/9-12 hr Range. Therefore, the screentime for study purpose has increased after the advent of the COVID-19.

***Analyzing Digital Tools preference for studying

```{r}
plot3= data.frame(study= df$bdevstudy, time= c("Before COVID-19"))
plot3= rbind(plot3, data.frame(study= df$adevstudy, time= c("After COVID-19")))
```
```{r}
plot3 %>%
  mutate(time = factor(time, levels = c("Before COVID-19","After COVID-19"))) %>%
  ggplot(aes(x=study, fill= time)) +
  geom_bar(position="dodge") +
  labs(title="Digital tools primarily used for studying?", x= "Response") + 
  theme_bw()
```

From the above barchart, it is very evident and clearly points out that students who might have used books or materials for studying have switched over to digital tools.

***Analyzing Concentration while the usage of devices

```{r}
plot4= data.frame(concentration= df$bconc, time= c("Before COVID-19"))
plot4= rbind(plot4, data.frame(concentration= df$aconc, time= c("After COVID-19")))
```
```{r}
plot4 %>%
  mutate(time = factor(time, levels = c("Before COVID-19","After COVID-19"))) %>%
  ggplot(aes(x=concentration, fill= time)) +
  geom_bar(position="dodge") +
  labs(title="Cannot Concentrate while studying with the use of devices?", x= "Response") + 
  theme_bw()
```
From the above barchart, it is clearly visible that the students who are using devices for studying are not able to fully concentrate and are often getting distracted.

***Analyzing The sleeping schedule

```{r}
plot5= data.frame(res= df$bbed, time= c("Before COVID-19"))
plot5= rbind(plot5, data.frame(res= df$abed, time= c("After COVID-19")))
```
```{r}
plot5 %>%
  mutate(time = factor(time, levels = c("Before COVID-19","After COVID-19"))) %>%
  ggplot(aes(x=res, fill= time)) +
  geom_bar(position="dodge") +
  labs(title="Fixed sleeping schedule?", x= "Response") + 
  theme_bw()
```
From the above barchart, we can tell that the students are not able to follow a fixed sleeping schedule and are not able to wake up on time after COVID-19 as the number of students who Disagree/Strongly disagree has risen in numbers. 

***Analyzing The Reason for sleep schedule

```{r}
plot6= data.frame(res= df$baffbed, time= c("Before COVID-19"))
plot6= rbind(plot6, data.frame(res= df$aaffbed, time= c("After COVID-19")))
```
```{r}
plot6 %>%
  mutate(time = factor(time, levels = c("Before COVID-19","After COVID-19"))) %>%
  ggplot(aes(x=res, fill= time)) +
  geom_bar(position="dodge") +
  labs(title="Devices responsible for affected sleeping schedule?", x= "Response") + 
  theme_bw()
```
From the above barchart, we can tell that the devices are one of the main reasons for the sleeping schedule of the students to be affected since the students who have strongly is very high and distinctly notable.

***Analyzing Tiredness due to screens

```{r}
plot7= data.frame(res= df$btire, time= c("Before COVID-19"))
plot7= rbind(plot7, data.frame(res= df$atire, time= c("After COVID-19")))
```
```{r}
plot7 %>%
  mutate(time = factor(time, levels = c("Before COVID-19","After COVID-19"))) %>%
  ggplot(aes(x=res, fill= time)) +
  geom_bar(position="dodge") +
  labs(title="Devices screentime resulting in Tiredness?", x= "Response") + 
  theme_bw()
```
From the above barchart, we can tell that the device screentime is one of the major reasons for resulting tiredness in students after COVID-19 since they need to be on their devices for both the classes as well as the assignments.

```{r}
write.csv(df, "C:/Users/rogha/OneDrive/Documents/MS WORD FILES/EDA/Project/Review-2/preprocess.csv", row.names=FALSE)
```
